Fix #24.
If an axis transition interrupts an earlier axis transition, the new transition
will see the exiting ticks for the old transition. If the axis is ordinal, the
position of these ticks may be undefined for both the old and new scale, which
would result in an invalid transform such as `translate(NaN,0)`.

If the new scale’s position for an exiting tick is undefined (as is always the
case with ordinal scales), rather than falling back to the old scale (which may
also be undefined), we simply use the tick’s current transform so that it fades
out in-place.